requirements.txt
matplotlib
msgpack
numpy
arrow
pybursts
ripe-atlas-cousteau
kafka-python
reverse_geocoder